Second part of default language -- REBASE when merging into master
It turns out that when you change the neutral language, the "(Default)"
localization is used if the system language matches it, even if there is
an explicit localization for the system language.

TL;DR the strings are no longer blank on English systems, and now as
intended English is used in the absence of localization.