Check native registry view for Windows version
This fixes Coverity CID 171806

On Windows the CurrentVersion key is only present in the native view
of the registry, so a 32-bit application checking the
windows version on 64-bit Windows needs to check the 64-bit registry
view explicitly.

In practice, the wintray app will never be running in 32-bit mode on
64-bit Windows because we let it run in AnyCPU mode.

The default behavior of checking the native registry view should have
been correct as a result, but it is technically possible for the app
to be run in x86 mode which is where the Coverity defect comes from.

This will make sure the code does the right thing either way.